python/using_flex_template_adv3/requirements.txt
apache-beam[gcp]==2.36.0
excel2json-3
